% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notion.R
\name{createPage}
\alias{createPage}
\title{Creates a new notion page from a parent page}
\usage{
createPage(basePage, pageTitle, notionKey)
}
\arguments{
\item{basePage}{The base page id.}

\item{pageTitle}{The title of the new page.}

\item{notionKey}{Secret notion key}
}
\value{
The page id
}
\description{
Creates a new notion page from a parent page
}
