% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notion.R
\name{updatePage}
\alias{updatePage}
\title{Updates a notion page with new content}
\usage{
updatePage(pageid, color, content, notionKey)
}
\arguments{
\item{pageid}{The page id}

\item{color}{The text colour}

\item{content}{The text to add to the page}

\item{notionKey}{Secret notion key}
}
\value{
A parsed response object
}
\description{
Updates a notion page with new content
}
